move test to RetriableStreamTest
